added refactoring validations to cancel a consultation
