+++ src/main/kotlin/com/jsrdev/med_api/domain/consult/validations/cancel/CancellationReason.kt
@@ -1,4 +1,4 @@
-package com.jsrdev.med_api.domain.consult.cancel
+package com.jsrdev.med_api.domain.consult.validations.cancel
 
 enum class CancellationReason {
     PATIENT_DESISTED,
